% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSR.r
\name{addSR}
\alias{addSR}
\title{addSR}
\usage{
addSR(lf, sr, leg = TRUE, hide = TRUE)
}
\arguments{
\item{lf}{\code{leaflet} object or proxy}

\item{sr}{\code{sf} object containing seasonal range polygons}

\item{leg}{logical - add a legend?}

\item{hide}{logical - hide seasonal range polygons on render?}
}
\value{
\code{leaflet} object
}
\description{
Add seasonal range polygons to a \code{leaflet} map
}
\examples{

sr <- wgfd_agol_data("Antelope_Seasonal_Range")

leaflet::leaflet() |>
  leaflet::addProviderTiles(leaflet::providers$Esri.WorldTopoMap) |>
  addSR(sr, hide = FALSE)

}
