% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_filter.R
\name{seasonal_filter}
\alias{seasonal_filter}
\title{Seasonal Filter}
\usage{
seasonal_filter(x, date_col, start, end)
}
\arguments{
\item{x}{data object to filter}

\item{date_col}{column to use for filtering (unquoted a la \code{dplyr})}

\item{start}{\code{Date} or \code{POSIX} start date in any year}

\item{end}{\code{Date} or \code{POSIX} end date in any year}
}
\value{
an object of the same class as \code{x}
}
\description{
Filters a data frame or similar object by date regardless of year
}
\details{
Start and end dates are inclusive, and filtering is performed on
julian day to correctly align filters with calendar dates. This means
one additional day may be included in leap years depending on whether
February 29 falls between the start and end dates specified. If
\code{start} and \code{end} are the same only records from that date
are returned. This function is timezone agnostic. Times are ignored
if \code{start} and/or \code{end} are provided as \code{POSIX} objects.
}
\examples{

library(tibble)

df <- tibble(
  name = letters[1:5],
  timestamp = as.POSIXct(c(
    "2020-01-01 08:00:00",
    "2022-10-05 11:15:00",
    "2023-12-10 19:30:00",
    "2024-08-31 00:00:00",
    "2024-05-25 02:45:00"
  )),
  x = 42 + runif(5, 0, 1),
  y = -107 + runif(5, 0, 1)
)

start <- as.Date("1900-10-01")
end <- as.Date("2030-05-01")

df |>
  seasonal_filter(timestamp, start, end)

}
