% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.utils-package.R
\docType{package}
\name{sr.utils-package}
\alias{sr.utils}
\alias{sr.utils-package}
\title{sr.utils: WGFD Seasonal Range Utilities}
\description{
Code for delineating wildlife seasonal range.
}
\keyword{internal}
