% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Utilities}
\alias{Utilities}
\alias{get_contour}
\alias{get_vol_ud}
\alias{is_vol_ud}
\title{Utilities}
\usage{
get_contour(raster_path, contour_level, crs = 4326)

get_vol_ud(r)

is_vol_ud(r)
}
\arguments{
\item{raster_path}{Path to raster file}

\item{contour_level}{Contour level to extract (0 - 1)}

\item{crs}{Coordinate reference system for output polygon}

\item{r}{Raster object}
}
\value{
\code{sf} polygon object

Raster object

logical
}
\description{
Basic functions for working with UD rasters
}
\details{
These might work better in a more generic package.
}
\section{Functions}{
\itemize{
\item \code{get_contour()}: Extract a contour polygon from a volume UD raster

\item \code{get_vol_ud()}: Convert a raster to a volume UD in which each cell
contains the CDF value corresponding to the value from the original raster

\item \code{is_vol_ud()}: Check the values of a raster to see if it has already
been converted to a volume UD. Simply checks if values range from 0 to 1.

}}
\examples{

ex_rast <- system.file("ex/elev.tif", package = "terra")

ct_90 <- get_contour(ex_rast, 0.9)

plot(ct_90)

r <- terra::rast(ex_rast)

is_vol_ud(r)

ud <- get_vol_ud(r)

is_vol_ud(ud)

terra::plot(ud)

}
